Split the 580-line DiagnosticQuestionForm into focused parts
The question authoring form was one 580-line component mixing metadata fields,
the stem, the options editor, and the diagram field. Split it via a shared
form context (FormProvider):

- questionForm/types.ts — form values type + pure helpers (defaultValues,
  labelForIndex, getCorrectOptionLabel, getDiagramSvg{Create,Update}) + the
  skill/difficulty constants
- questionForm/QuestionMetaFields.tsx — topic/status/skill/difficulty grid
- questionForm/QuestionOptionsEditor.tsx — options useFieldArray + add/remove/
  mark-correct (positional labels)
- questionForm/QuestionDiagramField.tsx — existing diagram / paste-SVG / upload

DiagnosticQuestionForm.tsx is now a 117-line composer that owns the form
instance, reset-on-record-change, and submit, and re-exports the pure helpers
so the create/edit pages and the test import unchanged. Behaviour and rendered
DOM are identical.
